+++ packages/database/src/migrator.ts
@@ -76,6 +76,7 @@ import * as outboxDlqAlertsEventUnique from './migrations/074_outbox_dlq_alerts_
 import * as agentContracts from './migrations/075_agent_contracts.js';
 import * as ownerActivationTokens from './migrations/076_owner_activation_tokens.js';
 import * as workerHeartbeats from './migrations/077_worker_heartbeats.js';
+import * as storefrontReadModelCache from './migrations/078_storefront_read_model_cache.js';
 import type { Database } from './types.js';
 
 const migrationProvider: MigrationProvider = {
@@ -158,6 +159,7 @@ const migrationProvider: MigrationProvider = {
       '075_agent_contracts': agentContracts,
       '076_owner_activation_tokens': ownerActivationTokens,
       '077_worker_heartbeats': workerHeartbeats,
+      '078_storefront_read_model_cache': storefrontReadModelCache,
     };
   },
 };